Hacked MQTT support into GoveeBTTempLogger
GoveeBTTempLogger seems to have the best/most complete govee
support that I've found of the online projects. I'd like to have
realtime updates of the latest data in things like homeassistant
in addtion to having it create the cool graphs it makes, so this is
a hack to cause the govee data to be updated via mqtt.

It's not clean or pretty in any way and hard codes connecting to the
local machine for the mqtt server.  But, it seems to work and is useful
to me for now.  A cleanup would be great.

See also my fork of govee-gateway
https://github.com/house-mouse/govee-gateway
which does mostly the same thing, taking the existing govee-gateway
code and putting the data parsers from GoveeBTTempLogger into it
to update mqtt.